+++ src/Event/ProgressAdvanceEvent.php
@@ -1,8 +1,8 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace MF\ConsoleSubscriber\Event;
 
-use Symfony\Component\EventDispatcher\Event;
+use Symfony\Contracts\EventDispatcher\Event;
 
 class ProgressAdvanceEvent extends Event
 {