+++ tests/AbstractTestCase.php
@@ -1,21 +1,18 @@
-<?php
+<?php declare(strict_types=1);
 
-namespace MF\Tests;
+namespace MF\ConsoleSubscriber;
 
-use Mockery as m;
+use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 use PHPUnit\Framework\TestCase;
 
 abstract class AbstractTestCase extends TestCase
 {
+    use MockeryPHPUnitIntegration;
+
     protected function expectsUnicode(string $withUnicode, string $withoutUnicode): string
     {
         return '\\' === DIRECTORY_SEPARATOR
             ? $withoutUnicode
             : $withUnicode;
-    }
-
-    protected function tearDown()
-    {
-        m::close();
     }
 }